ex05b/requirements.txt
python-dotenv
cassandra-driver
prettytable
PyYAML
cqlsh
